Do not assume a JobLogger receives a logger
In sidekiq/sidekiq#6200, the constructor for JobLogger was changed to
accept an entire `Sidekiq::Config` rather than a logger.

To maintain compatibility with previous versions of Sidekiq, the
constructor for Logcraft's Sidekiq::JobLogger has been changed to handle
both cases by checking the type of the argument before assigning the
logger.